udev: fixed udev limitation to start a long-running process
If an user script is started by udev and the script launches a long-running
process, then this process is terminated after around 2 minutes. The reason
is that udev (recent versions) prevents starting long-lived background
processes in RUN scripts. Hence this udev rule signals to systemd to start
such long-running process as a system service.