stm32: used a hardware timer (TIM3) to dim the LED display
The hardware timer 3 is configured to generate 2KHz software event (period
elapsed). After sending the pattern bytes to the LED board, the LED blocks
are not turned on directly if the dimmer flag is activated. Instead timer 3
is started. Once the timer 3 fires the period elapsed interrupt, the timer
operation is stopped, and the LED blocks are turned on.